Add files via upload
In this project, a power grid system where data of connected houses with different propagators are stored using a Python program with the help of graph data structures. The goal is to use graph data structure and perform certain operations such addition, and removal of houses and propagators from the designed structure. Here we are using a dictionary as an adjacency list to store the data in the graph.In this project, a power grid system where data of connected houses with different propagators are stored using a Python program with the help of graph data structures. The goal is to use graph data structure and perform certain operations such addition, and removal of houses and propagators from the designed structure. Here we are using a dictionary as an adjacency list to store the data in the graph.
Skills: Graphs